Lightrag/requirements.txt
lightrag-hku
neo4j
sentence-transformers
openai
python-dotenv
pytest
pypdf
python-docx
faiss-cpu
rich